initial commit for cleanup of hex drawing and clicking

+++ TurnRPG.Client/Utils/Point.cs
@@ -5,10 +5,10 @@ namespace TurnRPG.Client.Utils
     [Serializable]
     public class Point
     {
-        public int X { get; set; }
-        public int Y { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
 
-        public Point(int x, int y)
+        public Point(double x, double y)
         {
             X = x;
             Y = y;
